requirements.txt
aiosqlite==0.18.0
anyio==3.6.2
black==22.12.0
click==8.1.3
colorama==0.4.6
Deprecated==1.2.13
dnspython==2.2.1
docstring-parser==0.12
email-validator==1.3.0
fastapi==0.88.0
googleapis-common-protos==1.56.1
protobuf==3.20.3
grpcio
grpcio-tools
h11==0.14.0
idna==3.4
inflection==0.5.1
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
opentelemetry-api==1.15.0
opentelemetry-exporter-jaeger-proto-grpc==1.15.0
opentelemetry-instrumentation==0.36b0
opentelemetry-instrumentation-grpc==0.36b0
opentelemetry-sdk==1.15.0
opentelemetry-semantic-conventions==0.36b0
pathspec==0.10.3
piccolo==0.105.0
platformdirs==2.6.2
pydantic==1.10.4
sniffio==1.3.0
starlette==0.22.0
targ==0.3.7
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.20.0
wrapt==1.14.1
